requirements.txt
Django==3.1.2
django-allauth==0.42.0
django-crispy-forms==1.9.2
docker==4.3.1
gunicorn==20.0.4
prometheus-client==0.8.0
rfc3987==1.3.8
python-memcached==1.59
qrcode==6.1
django-widget-tweaks==1.4.8
pillow==7.2.0
